+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- chore: remove files endpoints in favor of uploads
+
 ## [0.9.1] - 2026-07-15
 
 ### Added